requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bitarray==2.9.2
cbor2==5.6.4
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==1.0.2
crytic-compile==0.3.7
cytoolz==0.12.3
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.8
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.1.0
packaging==24.1
parsimonious==0.10.0
prettytable==3.11.0
protobuf==5.28.0
pycryptodome==3.20.0
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.20.0
slither-analyzer==0.10.4
solc-select==1.0.4
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
web3==6.20.3
websockets==13.0.1
yarl==1.11.1
